back-end/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
openai
sentence-transformers
faiss-cpu
numpy
requests
pydantic